website_requirements.txt
flask==3.0.0
aiohttp==3.9.1
python-dotenv==1.0.0
requests==2.31.0
jinja2==3.1.2
gunicorn==21.2.0